working physicsshape with terrain LOD
The solution is ugly but it works.
Basically I must keep track of the havok index of each chunk myself, which means that any chunk deletion must check all other chunk index for update. What a time to be alive.

The chunktree + chunkforge system needs a rewrite it is difficult to keep track of chunks between the 2