Refactor CosmosJourneyer class
Instead of using an async init, the class will directly be created using a static async function.

This enables the removal of many null checks as everything will be defined when creating the engine instance.

This also allows to remove many bloat methods that were only useful for said null checks