Proof of concept
WebGPU can run both BabylonJS and LLM inference side by side

Now we wait for multimodal models for a complete space flight computer.

The goal of the flight computer will be to give answers to the player based on what is on screen, enriched with gameplay data.

Extensions are also possible for large actions model to control the spaceship by voice only, which would enable disabled persons to enjoy the cosmos.

I see a lot of potential.