Introduce StarSystemCoordinates
Developping the new universe object id system earlier showed some limits. Systems were identified by their seed.

The issue is that excludes custom star systems from interacting with the engine in the same way as seeded systems.

To fix this issue, star system coordinates are implemented by all types of star system models. It is easy to generate them from a seed and they can be hardcoded for custom systems.

This is future proof for the save file.

Right now, there is no good way to get the system model from the coordinates of a custom system, but a global registry could be used.