Rework rings model
Making the playground for saturn showed quite a large delta between the simulation and reality

This commit focuses on the definition of the rings, to use absolute distances instead of absolute. This makes defining the model for real objects much easier.

This commit also tweaks the density calculation for the rings to make them denser